requirements.txt
Django~=2.2.12
django-multiselectfield